Move hook to new file
