+++ client/my-sites/stats/feedback/use-on-screen.ts
@@ -1,0 +1,21 @@
+import { useEffect, useMemo, useState } from 'react';
+
+// Determines if an element is visible on screen or not.
+
+function useOnScreen( ref ) {
+	const [ isIntersecting, setIntersecting ] = useState( false );
+
+	const observer = useMemo(
+		() => new IntersectionObserver( ( [ entry ] ) => setIntersecting( entry.isIntersecting ) ),
+		[]
+	);
+
+	useEffect( () => {
+		observer.observe( ref.current );
+		return () => observer.disconnect();
+	}, [ observer, ref ] );
+
+	return isIntersecting;
+}
+
+export default useOnScreen;
